OCR Model Exploration/requirements.txt
Pillow==10.2.0
pytesseract==0.3.10
paddlepaddle==2.6.0 
paddleocr==2.7.3
requests==2.31.0 